03-3090:moving the useFormFilters into the FormsDashbaord component
